Switch 3d (well really 2.5d) to be a 2d graph
This commit switches the 3d graphs to be a much clearer to read 2d
hexbin graph. Doing this makes it much easier to see our data is
useless and doesn't show us anything except where we're bandwidth limited
either on the client or on firehose.openstack.org. We'll need to
reevaluate our benchmarking methodology to actually test the throughput
limits on the broker. (which was the intent of the benchmarking)